+++ fs/nfsd/nfs4acl.c
@@ -43,7 +43,8 @@
 #include "nfsd.h"
 #include "acl.h"
 #include "vfs.h"
-#include "nfs41acl_xdr.h"
+#include "../nfs_common/nfs41acl_xdr.h"
+#include <linux/nfsacl.h>	/* For convert_nfs41xdr_to_nfs40_acl and generate_nfs41acl_buf */
 
 #define NFS4_ACL_TYPE_DEFAULT	0x01
 #define NFS4_ACL_DIR		0x02
@@ -175,73 +176,6 @@ out:
 	posix_acl_release(dpacl);
 rel_pacl:
 	posix_acl_release(pacl);
-	return error;
-}
-
-static int
-convert_to_nfs40_ace(u32 *xdrbuf, struct nfs4_ace *ace)
-{
-	int error = 0;
-	u32 iflag, id;
-
-	ace->type = ntohl(*(xdrbuf++));
-	if (ace->type > NFS4_ACE_ACCESS_DENIED_ACE_TYPE)
-		return -EINVAL;
-
-	ace->flag = ntohl(*(xdrbuf++));
-	iflag = ntohl(*(xdrbuf++));
-	ace->access_mask = ntohl(*(xdrbuf++)) & NFS4_ACE_MASK_ALL;
-	id = ntohl(*(xdrbuf++));
-
-	if (iflag & ACEI4_SPECIAL_WHO) {
-		switch(id) {
-		case ACE4_SPECIAL_OWNER:
-			ace->whotype = NFS4_ACL_WHO_OWNER;
-			break;
-		case ACE4_SPECIAL_GROUP:
-			ace->whotype = NFS4_ACL_WHO_GROUP;
-			break;
-		case ACE4_SPECIAL_EVERYONE:
-			ace->whotype = NFS4_ACL_WHO_EVERYONE;
-			break;
-		}
-	} else {
-		ace->whotype = NFS4_ACL_WHO_NAMED;
-		if (ace->flag & NFS4_ACE_IDENTIFIER_GROUP) {
-			ace->who_gid = make_kgid(&init_user_ns, id);
-			if (!gid_valid(ace->who_gid)) {
-				error = -EINVAL;
-			}
-		} else {
-			ace->who_uid = make_kuid(&init_user_ns, id);
-			if (!uid_valid(ace->who_uid)) {
-				error = -EINVAL;
-			}
-		}
-	}
-
-	return error;
-}
-
-static int
-convert_nfs41xdr_to_nfs40_acl(u32 *xdrbuf, size_t remaining, struct nfs4_acl *acl)
-{
-	int error = 0;
-	int i;
-
-	for (i = 0; i < acl->naces; i++, xdrbuf += NACE41_LEN) {
-		if (remaining < (NACE41_LEN * sizeof(u32))) {
-			error = -EOVERFLOW;
-			break;
-		}
-
-		error = convert_to_nfs40_ace(xdrbuf, &acl->aces[i]);
-		if (error)
-			break;
-
-		remaining -= (NACE41_LEN * sizeof(u32));
-	}
-
 	return error;
 }
 
@@ -977,67 +911,6 @@ nfsd4_acl_to_attr_posix(enum nfs_ftype4 type, struct nfs4_acl *acl,
 		return nfserr_attrnotsupp;
 	else
 		return nfserrno(host_error);
-}
-
-static int
-convert_ace_to_nfs41(u32 *p, const struct nfs4_ace *ace)
-{
-	int error = 0;
-	u32 iflag = 0, who = -1;
-
-	/* Audit and Alarm are not currently supported */
-	if (ace->type > NFS4_ACE_ACCESS_DENIED_ACE_TYPE)
-		return -EINVAL;
-
-	switch (ace->whotype) {
-	case NFS4_ACL_WHO_OWNER:
-		iflag = ACEI4_SPECIAL_WHO;
-		who = ACE4_SPECIAL_OWNER;
-		break;
-	case NFS4_ACL_WHO_GROUP:
-		iflag = ACEI4_SPECIAL_WHO;
-		who = ACE4_SPECIAL_GROUP;
-		break;
-	case NFS4_ACL_WHO_EVERYONE:
-		iflag = ACEI4_SPECIAL_WHO;
-		who = ACE4_SPECIAL_EVERYONE;
-		break;
-	case NFS4_ACL_WHO_NAMED:
-		if (ace->flag & NFS4_ACE_IDENTIFIER_GROUP)
-			who = (u32)__kgid_val(ace->who_gid);
-		else
-			who = (u32)__kuid_val(ace->who_uid);
-		break;
-	default:
-		error = -EINVAL;
-	}
-
-	*p++ = htonl(ace->type);
-	*p++ = htonl(ace->flag & NFS41_FLAGS);
-	*p++ = htonl(iflag);
-	*p++ = htonl(ace->access_mask & NFS4_ACE_MASK_ALL);
-	*p++ = htonl(who);
-
-	return error;
-}
-
-static int
-generate_nfs41acl_buf(u32 *xdrbuf, const struct nfs4_acl *acl)
-{
-	int error = 0;
-	int i;
-
-	/* first byte is NFS41 Flags. Skip since these are RFC3530 acls */
-	*xdrbuf++ = 0;
-	*xdrbuf++ = htonl(acl->naces);
-
-	for (i = 0; i < acl->naces; i++, xdrbuf += NACE41_LEN) {
-		error = convert_ace_to_nfs41(xdrbuf, &acl->aces[i]);
-		if (error)
-			break;
-	}
-
-	return error;
 }
 
 static __be32
