+++ include/linux/nfsacl.h
@@ -11,6 +11,9 @@
 #include <linux/posix_acl.h>
 #include <linux/sunrpc/xdr.h>
 #include <uapi/linux/nfsacl.h>
+#if CONFIG_TRUENAS
+#include <linux/nfs4.h>		/* For struct nfs4_acl */
+#endif /* CONFIG_TRUENAS */
 
 /* Maximum number of ACL entries over NFS */
 #define NFS_ACL_MAX_ENTRIES	1024
@@ -44,5 +47,12 @@ nfs_stream_decode_acl(struct xdr_stream *xdr, unsigned int *aclcnt,
 extern bool
 nfs_stream_encode_acl(struct xdr_stream *xdr, struct inode *inode,
 		      struct posix_acl *acl, int encode_entries, int typeflag);
+
+#if CONFIG_TRUENAS
+extern int
+convert_nfs41xdr_to_nfs40_acl(u32 *xdrbuf, size_t remaining, struct nfs4_acl *acl);
+extern int
+generate_nfs41acl_buf(u32 *xdrbuf, const struct nfs4_acl *acl);
+#endif /* CONFIG_TRUENAS */
 
 #endif  /* __LINUX_NFSACL_H */
